src/kafka_etl/requirements.txt
kafka-python
pycouchdb
python-dotenv
requests
urllib3